# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/wallets": {
      "post": {
        "operationId": "WalletsController_create",
        "summary": "Create a new wallet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWalletDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The wallet has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "name": "My Bank Account",
                    "initial_balance": 1000,
                    "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "created_at": "2025-07-28T00:00:00.000Z",
                    "updated_at": "2025-07-28T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "WalletsController_findAll",
        "summary": "Get all wallets for the current user",
        "parameters": [
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of wallets with their current balances.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                      "name": "My Bank Account",
                      "initial_balance": 1000,
                      "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                      "created_at": "2025-07-28T00:00:00.000Z",
                      "updated_at": "2025-07-28T00:00:00.000Z",
                      "current_balance": 1250.5
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/wallets/{id}": {
      "get": {
        "operationId": "WalletsController_findOne",
        "summary": "Get a specific wallet by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The wallet with its current balance.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "name": "My Bank Account",
                    "initial_balance": 1000,
                    "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "created_at": "2025-07-28T00:00:00.000Z",
                    "updated_at": "2025-07-28T00:00:00.000Z",
                    "current_balance": 1250.5
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "WalletsController_update",
        "summary": "Update a wallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWalletDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated wallet.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "name": "My Updated Bank Account",
                    "initial_balance": 1500,
                    "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "created_at": "2025-07-28T00:00:00.000Z",
                    "updated_at": "2025-07-28T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "WalletsController_remove",
        "summary": "Delete a wallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The wallet has been successfully deleted."
          }
        },
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "summary": "Create a new category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "name": "Salary",
                    "type": "income",
                    "icon": "briefcase-outline",
                    "color": "#26de81",
                    "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "created_at": "2025-07-28T00:00:00.000Z",
                    "updated_at": "2025-07-28T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "summary": "Get all categories for the current user (including defaults)",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "income",
                "expense"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of categories.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                      "name": "Salary",
                      "type": "income",
                      "icon": "briefcase-outline",
                      "color": "#26de81",
                      "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                      "created_at": "2025-07-28T00:00:00.000Z",
                      "updated_at": "2025-07-28T00:00:00.000Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "summary": "Get a specific category by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The category.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "name": "Salary",
                    "type": "income",
                    "icon": "briefcase-outline",
                    "color": "#26de81",
                    "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "created_at": "2025-07-28T00:00:00.000Z",
                    "updated_at": "2025-07-28T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "summary": "Update a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated category.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "name": "Freelance Income",
                    "type": "income",
                    "icon": "laptop-outline",
                    "color": "#45aaf2",
                    "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                    "created_at": "2025-07-28T00:00:00.000Z",
                    "updated_at": "2025-07-28T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "summary": "Delete a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The category has been successfully deleted."
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Wallport API",
    "description": "API for Personal Finance Tracker Application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWalletDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Bank Account"
          },
          "initial_balance": {
            "type": "number",
            "example": 1000
          }
        },
        "required": [
          "name",
          "initial_balance"
        ]
      },
      "UpdateWalletDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Primary Bank Account"
          },
          "initial_balance": {
            "type": "number",
            "example": 1500
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Salary"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "income"
          },
          "icon": {
            "type": "string",
            "example": "briefcase-outline"
          },
          "color": {
            "type": "string",
            "example": "#26de81"
          }
        },
        "required": [
          "name",
          "type",
          "icon",
          "color"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Freelance Income"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "income"
          },
          "icon": {
            "type": "string",
            "example": "laptop-outline"
          },
          "color": {
            "type": "string",
            "example": "#45aaf2"
          }
        }
      },
...
    }
  }
}